docs(navigation): move deep linker docs to ionic page
